FifthMilestone/notifier/requirements.txt
Flask
flask_socketio
gunicorn
python-dotenv
redis
bcrypt
PyJWT
requests
APScheduler
eventlet